+++ cddagl/platform_api/utils.py
@@ -1,0 +1,5 @@
+import platform
+
+
+def is_windows():
+    return platform.system() == "Windows"
