+++ cddagl/posix_polyfill/pywintypes.py
@@ -1,0 +1,6 @@
+from cddagl.platform_api import PosixNotImplementedError
+
+
+class PyWinError(Exception):
+    def __int__(self):
+        raise PosixNotImplementedError()
