+++ cddagl/ui/views/settings.py
@@ -17,7 +17,7 @@ from cddagl.constants import get_locale_path, get_cdda_uld_path
 from cddagl.functions import clean_qt_path
 from cddagl.i18n import load_gettext_locale, get_available_locales, proxy_gettext as _
 from cddagl.sql.functions import get_config_value, set_config_value, config_true
-from cddagl.win32 import get_ui_locale
+from cddagl.platform_api import get_ui_locale
 
 logger = logging.getLogger('cddagl')
